+++ kipper/core/src/compiler/compiler.ts
@@ -337,7 +337,7 @@ export class KipperCompiler {
 	}
 
 	/**
-	 * Creates a new {@link KipperProgramContext} based on the passed {@link parseData} and {@link config configuration}.
+	 * Creates a new {@link KipperProgramContext} based on the passed {@link parseData} and {@link compilerOptions}.
 	 * @param parseData The parsing data of the file.
 	 * @param compilerOptions The compilation config.
 	 * @return The newly created {@link KipperProgramContext} instance, which contains the metadata of the compiled